requirements.txt
asgiref==3.6.0
certifi==2022.12.7
channels==4.0.0
chardet==5.1.0
charset-normalizer==3.0.1
Django==4.1.5
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.4.0